requirements.txt
PyYAML==6.0
google-cloud-bigquery==2.28.1